Sorted vms in separate folders
