+++ Garage_2_0/Models/ViewModels/Common/ErrorViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Garage_2_0.Models.ViewModels;
+﻿namespace Garage_2_0.Models.ViewModels.Common;
 
 
 public class ErrorViewModel
