+++ Garage_2_0/Models/ViewModels/Garage/GarageViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Garage_2_0.Models.ViewModels
+﻿namespace Garage_2_0.Models.ViewModels.Garage
 {
     public class GarageViewModel
     {
